Create tdoa_matlab file
This file contains the Matlab code for a collection of cooperative nodes (3 or more) to perform TDoA calculations against an uncooperative or unknown node.  It is assumed that each one of the distributed cooperative nodes have ideal, matching clocks and have passed their observations to a fusion center for the calculation.